Update README.md
A buzzer is an electronic device that makes a buzzing sound when electricity is supplied. It is commonly used in alarms, indicators, and electronic projects.
A buzzer changes electric signals into sound.It can be active (sounds by itself when powered) or passive (needs a signal to make sound)
It usually works with small voltages (like 3V, 5V, or 12V).
Buzzers are used in doorbells, alarms, timers, and toys.